+++ src/index.ts
@@ -14,5 +14,6 @@ export {
 } from './instance';
 
 export { default as BaseClient } from './base_client';
+export { default as SDKInfo } from './sdk_info';
 
 export { TokenWithExpiry } from './authenticator';